Modify biothings jsonschema namespace (#370)
* Move the modules to be co-located where used ...

While the data schema generation may be fairly general it is only used in the
data inspection module. Attempting to co-locate modules to more
appropriate hub locations to avoid having utils be a polluted namespace

* Update the namespaces so we don't collide python builtin's

* Move the tests from the module to tests

* Combine the inspection and schema tests

* Change the import calls for cli inspect command

* Cleanup the dataplugin imports

* Rename inspect -> doc_inspect to avoid builtin shadowing

---------
