+++ tests/hub/datainspect/schemas/cgi_map.json
