+++ tests/hub/datainspect/schemas/cgi_schema.json
